agents/pyproject.toml
[project]
name = "archive-agents"
version = "0.1.0"
description = "Archive Agents - Decentralized Multi-Agent Protocol on NeoX using SpoonOS"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "spoon-ai-sdk>=0.1.0",
    "spoon-toolkits>=0.1.0",
    "web3>=7.0.0",
    "eth-account>=0.13.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "aiohttp>=3.10.0",
    "pydantic>=2.9.0",
    "twilio>=9.3.0",
    "uvicorn>=0.32.0",
    "fastapi>=0.115.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.0.0",
    "ruff>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
